Fixed empty list issue.
Using the given `.tab` and `fasta` datasets, the chromosomes were represented as `chr{chromosome}`, for example `chr1` or `chrX`. In the `seq` list the chromosomes were represented as `1`, `2`, ..., `X`,`Y`, so when the `intersect` function between the two sets was called, the result was an empty list. As solution is proposed to prepend `chr` to the chromosome names in the `seq` list.